Closes #37 and #38

+++ src/me/pheasn/blockown/owning/SQLOwningNetwork.java
@@ -5,11 +5,11 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import me.pheasn.OfflineUser;
 import me.pheasn.blockown.BlockOwn;
 import me.pheasn.blockown.mysql.MySqlNetwork;
 import me.pheasn.blockown.mysql.TableDefinition;
 
-import org.bukkit.OfflinePlayer;
 import org.bukkit.WorldCreator;
 import org.bukkit.block.Block;
 
@@ -26,11 +26,8 @@ public class SQLOwningNetwork extends SQLOwning {
 
 	@Override
 	public boolean load() {
-		return (msql
-				.connect(
-						plugin.getConfig().getString(
-								Setting.MYSQL_HOST.toString())
-								+ ":" + plugin.getConfig().getInt(Setting.MYSQL_PORT.toString()) + "/" + plugin.getConfig().getString(Setting.MYSQL_DATABASE.toString()), plugin.getConfig().getString(Setting.MYSQL_USER.toString()), plugin.getConfig().getString(Setting.MYSQL_PASSWORD.toString())) && createTablesIfNotExist()); //$NON-NLS-1$ //$NON-NLS-2$
+		return (msql.connect(plugin.getConfig().getString(
+				Setting.MYSQL_HOST.toString()) + ":" + plugin.getConfig().getInt(Setting.MYSQL_PORT.toString()) + "/" + plugin.getConfig().getString(Setting.MYSQL_DATABASE.toString()), plugin.getConfig().getString(Setting.MYSQL_USER.toString()), plugin.getConfig().getString(Setting.MYSQL_PASSWORD.toString())) && createTablesIfNotExist()); //$NON-NLS-1$ //$NON-NLS-2$
 	}
 
 	private boolean createTablesIfNotExist() {
@@ -43,17 +40,15 @@ public class SQLOwningNetwork extends SQLOwning {
 	}
 
 	@Override
-	public OfflinePlayer getOwner(Block block) {
+	public OfflineUser getOwner(Block block) {
 		String world = block.getWorld().getName();
 		int x = block.getX();
 		int y = block.getY();
 		int z = block.getZ();
-		ResultSet rs = msql
-				.doQuery("SELECT playername FROM block INNER JOIN player ON block.ownerid=player.playerid WHERE x=" + x + " AND y=" + y + " AND z=" + z + " AND world='" + world + "';"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
+		ResultSet rs = msql.doQuery("SELECT playername FROM block INNER JOIN player ON block.ownerid=player.playerid WHERE x=" + x + " AND y=" + y + " AND z=" + z + " AND world='" + world + "';"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
 		try {
 			if (rs.next()) {
-				return plugin.getServer().getOfflinePlayer(
-						rs.getString("playername")); //$NON-NLS-1$
+				return OfflineUser.getInstance(rs.getString("playername")); //$NON-NLS-1$
 			} else {
 				return null;
 			}
@@ -66,29 +61,22 @@ public class SQLOwningNetwork extends SQLOwning {
 	}
 
 	@Override
-	public void setOwner(Block block, OfflinePlayer offlinePlayer) {
-		if (offlinePlayer != null) {
-			setOwner(block, offlinePlayer.getName());
-		}
-	}
-
-	@Override
-	public void setOwner(Block block, String player) {
-		if (!playerExists(player)) {
-			msql.doUpdate("INSERT INTO player(playername) VALUES('" + player + "');"); //$NON-NLS-1$ //$NON-NLS-2$
+	public void setOwner(Block block, OfflineUser player) {
+		if (!playerExists(player.getName())) {
+			msql.doUpdate("INSERT INTO player(playername) VALUES('" + player.getName() + "');"); //$NON-NLS-1$ //$NON-NLS-2$
 		}
 		String world = block.getWorld().getName();
 		int x = block.getX();
 		int y = block.getY();
 		int z = block.getZ();
-		ResultSet rs = msql
-				.doQuery("SELECT playerid FROM player WHERE playername='" + player + "';"); //$NON-NLS-1$ //$NON-NLS-2$
+		ResultSet rs = msql.doQuery("SELECT playerid FROM player WHERE playername='" + player.getName() + "';"); //$NON-NLS-1$ //$NON-NLS-2$
 		int playerid = 0;
 		try {
 			if (rs.next()) {
 				playerid = rs.getInt("playerid"); //$NON-NLS-1$
 			}
 		} catch (SQLException e) {
+			plugin.error(e);
 		} catch (NullPointerException e) {
 		}
 		msql.doUpdate("INSERT INTO block(world, x, y, z, ownerid) VALUES('" + world + "', '" + x + "', '" + y + "', '" + z + "', '" + playerid + "') ON DUPLICATE KEY UPDATE world=VALUES(world), x=VALUES(x), y=VALUES(y), z=VALUES(z), ownerid=VALUES(ownerid);"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$ //$NON-NLS-6$
@@ -109,7 +97,7 @@ public class SQLOwningNetwork extends SQLOwning {
 	}
 
 	@Override
-	boolean playerExists(String player) {
+	protected boolean playerExists(String player) {
 		player = plugin.getServer().getOfflinePlayer(player).getName();
 		ResultSet rs = msql.doQuery("SELECT playername FROM player;"); //$NON-NLS-1$
 		ArrayList<String> players = new ArrayList<String>();
@@ -130,27 +118,21 @@ public class SQLOwningNetwork extends SQLOwning {
 	}
 
 	@Override
-	public void deleteOwningsOf(String player) {
-		ResultSet rs = msql
-				.doQuery("SELECT playerid FROM player WHERE playername='" + player + "';"); //$NON-NLS-1$ //$NON-NLS-2$
+	public void deleteOwningsOf(OfflineUser player) {
+		ResultSet rs = msql.doQuery("SELECT playerid FROM player WHERE playername='" + player.getName() + "';"); //$NON-NLS-1$ //$NON-NLS-2$
 		try {
-			rs.next();
+			if(rs.next()){
 			int playerid = rs.getInt("playerid"); //$NON-NLS-1$
 			msql.doUpdate("DELETE FROM block WHERE ownerid=" + playerid + ";"); //$NON-NLS-1$ //$NON-NLS-2$
+			}
 		} catch (SQLException e) {
 		}
 	}
 
 	@Override
-	public void deleteOwningsOf(OfflinePlayer offlinePlayer) {
-		deleteOwningsOf(offlinePlayer.getName());
-	}
-
-	@Override
-	public HashMap<Block, String> getOwnings() {
-		HashMap<Block, String> result = new HashMap<Block, String>();
-		ResultSet rs = msql
-				.doQuery("SELECT * FROM block INNER JOIN player ON block.ownerid=player.playerid;"); //$NON-NLS-1$
+	public HashMap<Block, OfflineUser> getOwnings() {
+		HashMap<Block, OfflineUser> result = new HashMap<Block, OfflineUser>();
+		ResultSet rs = msql.doQuery("SELECT * FROM block INNER JOIN player ON block.ownerid=player.playerid;"); //$NON-NLS-1$
 		try {
 			while (rs.next()) {
 				String world = rs.getString("world"); //$NON-NLS-1$
@@ -160,14 +142,12 @@ public class SQLOwningNetwork extends SQLOwning {
 				if (plugin.getServer().getWorld(world) == null) {
 					plugin.getServer().createWorld(new WorldCreator(world));
 				}
-				result.put(
-						plugin.getServer().getWorld(world).getBlockAt(x, y, z),
-						rs.getString("playername")); //$NON-NLS-1$
+				result.put(plugin.getServer().getWorld(world).getBlockAt(x, y, z), OfflineUser.getInstance(rs.getString("playername"))); //$NON-NLS-1$
 			}
 			rs.close();
 			return result;
 		} catch (SQLException e) {
-			return new HashMap<Block, String>();
+			return new HashMap<Block, OfflineUser>();
 		}
 	}
 }
